+++ src/components/Lottie/index.ts
@@ -1,2 +1,0 @@
-export { ErrorAnimation } from "./ErrorAnimation";
-export { SuccessAnimation } from "./SuccessAnimation";